\name{fitSimPoi}
\alias{fitSimPoi}
\docType{data}
\title{
	Fitted glmssn object for simulated data
}
\description{
  This is a fitted model using the \code{\link{glmssn}} function after simulating
	data.
}
\details{
	See the help for \code{\link{SimulateOnSSN}} for how the model was created.
}

\examples{
	library(SSN)
	data(modelFits)
	ls()
}
